PR Feedback - reorganize logic and add more comments
